+++ tests/test_api.py
@@ -1,4 +1,10 @@
+from typing import Optional
+
+import pytest
+from _pytest.monkeypatch import MonkeyPatch
+
 import platformdirs
+from platformdirs.android import Android
 
 
 def test_package_metadata() -> None:
@@ -16,3 +22,19 @@ def test_property_result_is_str(func: str) -> None:
     dirs = platformdirs.PlatformDirs("MyApp", "MyCompany", version="1.0")
     result = getattr(dirs, func)
     assert isinstance(result, str)
+
+
+@pytest.mark.parametrize("root", ["A", "/system", None])
+@pytest.mark.parametrize("data", ["D", "/data", None])
+def test_android_active(monkeypatch: MonkeyPatch, root: Optional[str], data: Optional[str]) -> None:
+    for env_var, value in {"ANDROID_DATA": data, "ANDROID_ROOT": root}.items():
+        if value is None:
+            monkeypatch.delenv(env_var, raising=False)
+        else:
+            monkeypatch.setenv(env_var, value)
+
+    expected = root == "/system" and data == "/data"
+    if expected:
+        assert platformdirs._set_platform_dir_class() is Android
+    else:
+        assert platformdirs._set_platform_dir_class() is not Android
